requirements.txt
pip
torch
imageio
gradio
tensorflow
Flask
gunicorn
Pillow
opencv_python
algorithmia
tf_slim
PyYaml
